pyproject.toml
[project]
name = "langchain-crash-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.0,<3.12"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.0.20",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.7",
    "langchain-redis>=0.2.3",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "sentence-transformers>=5.1.0",
    "tavily-python>=0.7.11",
    "tiktoken>=0.11.0",
    "wikipedia>=1.4.0",
]
